% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_locusplot.R
\name{fast_locusplot}
\alias{fast_locusplot}
\title{Function to generate zoomed locus plots including LD}
\usage{
fast_locusplot(data, build = "hg19", snp = "top", token = NULL)
}
\arguments{
\item{data}{A data.frame. Columns 'chr', 'pos', and 'pvalue' are required}

\item{build}{Genomic build. Currently only hg19 is supported.}

\item{snp}{Default top snp with lowest pvalue or rsid}

\item{token}{!!!REQUIRED!!!: Register for token at https://ldlink.nci.nih.gov/?tab=apiaccess}
}
\value{
A ggplot2 object/plot
}
\description{
Function to generate zoomed locus plots including LD
}
\examples{
fast_locusplot(gwas_data, token = "mytoken")
}
