% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_manhattan.R
\name{fast_manhattan}
\alias{fast_manhattan}
\title{A fast function for Manhattan plots}
\usage{
fast_manhattan(
  data,
  build = "hg19",
  color1 = "black",
  color2 = "grey",
  y_scale = TRUE,
  log10p = TRUE,
  alpha = 1,
  speed = "fast",
  pointsize = 0,
  pixels = c(512, 512),
  ...
)
}
\arguments{
\item{data}{Data have to contain the columns "chr", "pos" and "pvalue". Of note, chromosomes should be written with "chr"-tag like "chr1".}  
  
\item{build}{Genomic build. Currently hg18, hg19, and hg38 are supported. If you are not sure about, set the default "hg19".}

\item{color1}{Color for odd-numbered chromosomes}

\item{color2}{Color for even-numbered chromosomes}

\item{y_scale}{When an own y-scale is provided then set this to TRUE to avoid an error.}

\item{log10p}{If TRUE (default), the -log10 transformed pvalues are plotted}

\item{speed}{The speed option. Fast and ultrafast use \code{\link{scattermore}} functionality}

\item{pointsize}{Only when using the 'fast' option you can increase pointsize. Default is 0. When using small pointsizes it could be that points are not shown in the RStudio Plots or Zoom window. But they will plotted when saving to pdf.}

\item{pixels}{Only when using the 'fast' option you can increase pixel width and height. Default is c(512, 512).}

\item{gwas}{A data.frame. Columns "chr", "pos", and "pvalue" are required}
}
\value{
A ggplot2 object/plot
}
\description{
This is a very fast and easy-to-individualize plotting function for GWAS results e.g. pvalues based on \code{\link{ggplot2}} and \code{\link{scattermore}}.
}
\examples{
data("gwas_data")
head(gwas_data)
# slow
fast_manhattan(gwas_data, build='hg18')
# fast
fast_manhattan(gwas_data, build = "hg18", speed = "fast")
# ultrafast
fast_manhattan(gwas_data, build='hg18', speed = "ultrafast")
}
