% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_qq.R
\name{fast_qq}
\alias{fast_qq}
\title{A fast version of a quantile-quantile plot}
\usage{
fast_qq(
  pvalue,
  speed = "fast",
  pointsize = 1.2,
  linecolor = "deeppink",
  ci_color = "steelblue",
  ci_alpha = 0.3,
  conf.alpha = 0.05,
  confint = T,
  log10 = T,
  inflation_method = "median",
  title = "QQ-plot"
)
}
\arguments{
\item{pvalue}{A numeric vector of pvalues.}

\item{speed}{The speed option. Fast and ultrafast use scattermore functionality}

\item{pointsize}{Only when using the 'fast' option you can increase pointsize. Default is 1.2}

\item{linecolor}{Color of the abline of the theoretical and sample distributions.}

\item{ci_color}{Color fof the confidence interval}

\item{ci_alpha}{Alpha value (e.g. transparency) of the confidence's fill color}

\item{conf.alpha}{Alpha level of the confidence interval. Default value = 0.05}

\item{confint}{Print a confident interval. Default = TRUE}

\item{log10}{Should the pvalues -log10 tranformed? Default = TRUE}

\item{inflation_method}{Method to calculate the genomic inflation factor. choices = c("median","regression"). The "median" method is the most common way,while the regression method to estimate lambda is adapted from the GenABEL package.}

\item{title}{A title added to the plot}

\item{color2}{Fill color for even-numbered chromosomes}
}
\value{
A ggplot2 object/plot
}
\description{
This is a very fast plotting function for quantile-quantile (QQ) plots using \code{\link{ggplot2}} and \code{\link{scattermore}} functionalities.
}
\examples{
data("gwas_data")
head(gwas_data)
qq_fast(gwas_data$pvalue, inflation_method = "reg")
qq_fast(gwas_data$pvalue, inflation_method = "med")
}
