% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chrom_lengths.R
\name{get_chrom_lengths}
\alias{get_chrom_lengths}
\title{Get chromosome length depending on genome build}
\usage{
get_chrom_lengths(build = c("hg18", "hg19", "hg38"))
}
\description{
This function gives you chromosome lengths of implemented genome builds hg18,hg19, and hg38.
}
\examples{
get_chrom_lengths(build='hg19')
}
