% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Add together two numbers}
\usage{
manhattan(
  gwas,
  build = c("hg18", "hg19", "hg38"),
  color1 = "black",
  color2 = "grey",
  y_scale = TRUE,
  log10p = TRUE,
  alpha = 1,
  speed = c("slow", "fast", "ultrafast"),
  pointsize = 0,
  pixels = c(512, 512),
  ...
)
}
\arguments{
\item{gwas}{A data.frame. Columns "chr", "pos", and "pvalue" are required}

\item{build}{Genomic build. Currently hg18, hg19, and hg38 are supported}

\item{color1}{Color for odd-numbered chromosomes}

\item{color2}{Color for even-numbered chromosomes}

\item{y_scale}{When an own y-scale is provided then set this to TRUE to avoid an error.}

\item{log10p}{If TRUE (default), the -log10 transformed pvalues are plotted}

\item{speed}{The speed option. Fast and ultrafast use scattermore functionality}

\item{pointsize}{Only when using the 'fast' option you can increase pointsize. Default is 0. When using small pointsizes it could be that points are not shown in the RStudio Plots or Zoom window. But they will plotted when saving to pdf.  it}

\item{pixels}{Only when using the 'fast' option you can increase pixel width and height. Default is c(512, 512).}
}
\value{
A ggplot object that makes a Manhattan plot.
}
\description{
Add together two numbers
}
\examples{
data(cad_gwas)
cad_gwas$y=-log10(cad_gwas$pval)
manhattan(cad_gwas,build='hg18')
}
