% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Add together two numbers}
\usage{
manhattan(
  gwas,
  build = c("hg18", "hg19", "hg38"),
  color1 = "black",
  color2 = "grey",
  y_scale = TRUE,
  speed = c("slow", "fast", "ultrafast"),
  pointsize = 0,
  pixels = c(512, 512),
  ...
)
}
\arguments{
\item{gwas}{A GWAS study. Must have chrom, pos, and a y column reflecting -log10 pvalues}

\item{build}{Genomic build. Currently supports hg18, hg19, and hg38}

\item{color1}{Color for odd-numbered chromosomes}

\item{color2}{Color for even-numbered chromosomes}
}
\value{
Ta ggplot object that makes a Manhattan plot.
}
\description{
Add together two numbers
}
\examples{
data(cad_gwas)
cad_gwas$y=-log10(cad_gwas$pval)
manhattan(cad_gwas,build='hg18')
}
